Adds test for subscriptions
